+++ server/queues/tasks/DocumentAddGroupNotificationsTask.ts
@@ -1,6 +1,4 @@
 import { Op } from "sequelize";
-import { DocumentPermission } from "@shared/types";
-import Logger from "@server/logging/Logger";
 import { GroupUser } from "@server/models";
 import { DocumentGroupEvent } from "@server/types";
 import BaseTask, { TaskPriority } from "./BaseTask";
@@ -8,21 +6,6 @@ import DocumentAddUserNotificationsTask from "./DocumentAddUserNotificationsTask
 
 export default class DocumentAddGroupNotificationsTask extends BaseTask<DocumentGroupEvent> {
   public async perform(event: DocumentGroupEvent) {
-    const permission = event.changes?.attributes.permission as
-      | DocumentPermission
-      | undefined;
-
-    if (!permission) {
-      Logger.debug(
-        "task",
-        `Suppressing notification for group ${event.modelId} as permission not available`,
-        {
-          documentId: event.documentId,
-          groupId: event.modelId,
-        }
-      );
-    }
-
     await GroupUser.findAllInBatches<GroupUser>(
       {
         where: {
